Add containers to each box in html and customize css
Added seperate containers to each box in the html so I could customize the
boxes individually then combine them together. In css I gave each box
a width of 150px and a height of 150px then a border-color. Spaced the
text below the container by 5px to give it some room.